[Python dopar] Script for running commands in parallel - two separate processes.
Works with Python >= 2.4 (subprocess). To use it issue any command, which can 
process multiple arguments (can be also combined with xargs, if neccesary),
to run 2 separate copies, each one with a half of the arguments provided 
after the triple slash (///).

Just copy or symlink it somewhere on the $PATH.

Example:

dopar mogrify -resize 1024x1024 -verbose /// *.jpg